LayoutTests/fast/events/keyboardevent-location-constants.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.w3.org/TR/DOM-Level-3-Events/#events-KeyboardEvent" />
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Tests that the DOM_KEY_LOCATION_* constants are defined on KeyboardEvent");

shouldBe('KeyboardEvent.DOM_KEY_LOCATION_STANDARD', '0x00');
shouldBe('KeyboardEvent.DOM_KEY_LOCATION_LEFT', '0x01');
shouldBe('KeyboardEvent.DOM_KEY_LOCATION_RIGHT', '0x02');
shouldBe('KeyboardEvent.DOM_KEY_LOCATION_NUMPAD', '0x03');
// These constants are not yet supported (crbug.com/265446)
shouldBeUndefined('KeyboardEvent.DOM_KEY_LOCATION_MOBILE');
shouldBeUndefined('KeyboardEvent.DOM_KEY_LOCATION_JOYSTICK');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
